Changed animation location to main pages not components; Changed animation style; changed hasrouter location; Changed css zindex location;
