+++ src/App.js
@@ -1,21 +1,23 @@
 import React from 'react'
-import Navbar from './components/navbar/Navbar.js';
-import { HashRouter } from 'react-router-dom';
-import { Routes, Route } from 'react-router-dom';
+//import Navbar from './components/navbar/Navbar.js';
+
+import { Routes, Route, useLocation } from 'react-router-dom';
 import Home from './components/pages/Home.js'
 import Projects from './components/pages/Projects';
 import AboutMe from './components/pages/AboutMe';
 import Skills from './components/pages/Skills';
 import IndividualProject from './components/pages/IndividualProject.js';
 import TechnicalProjects from './components/pages/TechnicalProjects.js';
+import { AnimatePresence } from 'framer-motion';
 import './App.css'
 function App() {
 
+  const location = useLocation();
+
   return (
     <>
-      <HashRouter hashType="hashbang">
-        
-        <Routes>
+      <AnimatePresence mode='wait'>
+        <Routes location={location} key={location.key}>
           <Route path='/' exact Component={Home} />
           <Route path='/Personal-Website' exact Component={Home} />
           <Route path='/Projects' excat Component={Projects} />
@@ -24,7 +26,7 @@ function App() {
           <Route path='/IndividualProject' excat Component={IndividualProject} />
           <Route path='/TechnicalProjects' excat Component={TechnicalProjects} />
         </Routes>
-      </HashRouter>
+      </AnimatePresence>
     </>
   );
 }
