+++ src/components/HomeBody.css
@@ -3,6 +3,8 @@
     justify-content: center;
     padding-bottom: 60px;
     margin-top: 13px;
+    position: relative;
+    z-index: 0;
 }
 
 @media screen and (max-width: 600px) {
