+++ src/components/HomeBody.js
@@ -1,33 +1,17 @@
 import React from 'react'
 import HomeTitleCard from './HomeTitleCard'
 import './HomeBody.css'
-import { delay, motion } from 'framer-motion'
 
 function HomeBody() {
-
-    const childVariants = {
-        hidden: {
-            opacity: 0
-        },
-
-        visible: {
-            opacity: 1,
-            transition: {
-                delay: 1.0,
-                duration: 1
-            }
-        }
-
-    }
 
 
     return (
         <>
-            <motion.div className='HomeTitleCardContainer' variants={childVariants} initial="hidden" animate="visible">
+            <div className='HomeTitleCardContainer'>
                 <HomeTitleCard Title={"Projects"} link={"/Projects"} body={"Click the tile to view some of my projects I created over the years."}></HomeTitleCard>
                 <HomeTitleCard Title={"About Me"} link={"/AboutMe"} body={"Click the tile to see who I am, what education I have, and where do I work."}></HomeTitleCard>
                 <HomeTitleCard Title={"Skills"} link={"/Skills"} body={"Click the tile to see what skills I have acquired over the years."}></HomeTitleCard>
-            </motion.div>
+            </div>
         </>
     )
 }
