+++ src/components/HomeHeader.js
@@ -2,29 +2,11 @@ import React from 'react'
 import ButtonLinks from './buttonLinks/ButtonLinks'
 import './HomeHeader.css'
 import Navbar from './navbar/Navbar'
-import { motion } from 'framer-motion'
-
-const containerVarients = {
-    hidden: {
-        y: -850,
-    },
-
-    visible: {
-        y: 0,
-        transition: {
-            duration: .75,
-            type: 'spring',
-            stiffness: 60,
-           // when:"beforeChildren"
-        }
-    }
-}
-
-
+//import { motion } from 'framer-motion'
 
 function Header() {
     return (
-        <motion.div className="Header" variants={containerVarients} initial="hidden" animate="visible">
+        <div className="Header">
             <Navbar></Navbar>
             <h1 className="title-h1">Scott Imandt</h1>
             <h2 className="title-h2">Computer Science Major</h2>
@@ -52,7 +34,7 @@ function Header() {
                     />
                 </div>
             </div>
-        </motion.div>
+        </div>
     )
 }
