+++ src/components/IndividualHeader.css
@@ -10,6 +10,8 @@
     margin: auto;
     justify-content: center;
     align-items: center;
+    position:relative;
+    z-index: 999;
 }
 
 .title-h1 {
@@ -20,7 +22,8 @@
 }
 
 .maincontent {
-    max-width: 700px;
+    max-width: 1000px;
+    
 }
 
 .mainimg img {
