+++ src/components/PageCard.css
@@ -13,6 +13,8 @@
     padding-right: 50px;
     padding-top: 20px;
     padding-bottom: 50px;
+    position:relative;
+    z-index: 0;
 
 }
