+++ src/components/SkillsBody.js
@@ -1,28 +1,13 @@
 import React from 'react'
 import PageCard from './PageCard'
-import { motion } from 'framer-motion'
 
 function SkillsBody() {
 
-    const childVariants = {
-        hidden: {
-            opacity: 0
-        },
-
-        visible: {
-            opacity: 1,
-            transition: {
-                delay: 1.0,
-                duration: 1
-            }
-        }
-
-    }
 
     const myStyle = { marginRight: "7.5%", marginLeft: "7.5%", marginTop: "50px" };
 
     return (
-        <motion.div variants={childVariants} initial="hidden" animate="visible">
+        <>
             <PageCard
                 title={'Software Engineering'}
                 body={''}
@@ -66,7 +51,7 @@ function SkillsBody() {
             </div>
 
 
-        </motion.div>
+        </>
     )
 }
