+++ src/components/buttonLinks/ButtonLinks.css
@@ -5,8 +5,8 @@
     box-shadow: 0px 0px 20px 5px rgba(0,0,0,0.50);
     padding: 15px;
     width: 300px;
-    margin-left: 40px;
-    margin-right: 40px;
+    margin-left: 29px;
+    margin-right: 29px;
     border-radius: 50px;
 }
