+++ src/components/navbar/Navbar.css
@@ -12,6 +12,7 @@
     padding-bottom: 15px;
     padding-top:20px;
     font-family: 'Lobster', 'Merriweather Sans', sans-serif;
+    width: 100%;
   }
   
   .navbar-container {
