+++ src/components/pages/IndividualProject.js
@@ -3,6 +3,7 @@ import { useLocation } from 'react-router-dom'
 import IndividualHeader from '../IndividualHeader';
 import PageCard from '../PageCard';
 import PageArrayList from '../PageArrayList';
+import { motion } from 'framer-motion'
 
 function LoadJsonData() {
   const location = useLocation();
@@ -35,7 +36,7 @@ function LoadJsonData() {
 
     case "DIYTaskTracker":
       return require('../../data/Projectdata/DIYTaskTracker.json')
-    
+
     case "CICaidA: Custom Hardware":
       return require('../../data/Projectdata/CICaidACustom.json')
 
@@ -68,39 +69,97 @@ function IndividualProject() {
     )
   }
 
+  const containerVarients = {
+    hidden: {
+        y: -550,
+
+    },
+
+    visible: {
+        y: 0,
+
+        transition: {
+            duration: .75,
+            type: 'spring',
+            stiffness: 60,
+            //when: "beforeChildren"
+        }
+    },
+    exit: {
+        y: -550,
+        transition: {
+            type: 'tween',
+            duration: .3,
+            delay: .20
+        }
+
+    }
+}
+
+const childVariants = {
+    hidden: {
+        opacity: 0,
+        y: -300,
+
+    },
+
+    visible: {
+        y: 0,
+        opacity: 1,
+        transition: {
+            delay: .20,
+            duration: 1,
+            type: 'spring',
+            stiffness: 60,
+        }
+    },
+    exit: {
+        opacity: 0,
+        y: -1000,
+        transition: {
+            duration: .3,
+            type: 'tween',
+
+        }
+    }
+
+}
+
 
 
   // if data exsits on the website
   return (
     <>
+      <motion.div variants={containerVarients} initial="hidden" animate="visible" exit='exit'>
+        <IndividualHeader
+          projectName={data.title}
+          subtitle={data.subtitle}
+          mainImage={data.mainimgsrc}
+          githublink={data.githublink}
+        />
 
-      <IndividualHeader
-        projectName={data.title}
-        subtitle={data.subtitle}
-        mainImage={data.mainimgsrc}
-        githublink={data.githublink}
-      />
+        <motion.div variants={childVariants}>
+          <PageCard
+            title={'Overview'}
+            body={data.Overview}
+            img={""}>
+          </PageCard>
 
-      <PageCard
-        title={'Overview'}
-        body={data.Overview}
-        img={""}>
-      </PageCard>
+          <PageCard
+            title={'Objective'}
+            body={data.Objective}
+            img={""}>
+          </PageCard>
 
-      <PageCard
-        title={'Objective'}
-        body={data.Objective}
-        img={""}>
-      </PageCard>
+          <PageArrayList pageArrayList={pageArrayList}></PageArrayList>
 
-      <PageArrayList pageArrayList={pageArrayList}></PageArrayList>
-
-      <PageCard
-        title={'Conclusion'}
-        body={data.Conclusion}
-        img={""}>
-      </PageCard>
-
+          <PageCard
+            title={'Conclusion'}
+            body={data.Conclusion}
+            img={""}>
+          </PageCard>
+        </motion.div>
+      </motion.div>
     </>
   )
 }
