+++ src/index.js
@@ -1,11 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import { HashRouter } from 'react-router-dom';
 import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  
+    
+    <HashRouter hashType="hashbang">
     <App />
-  
+    </HashRouter>
 );
